Create Cars models

+++ TrafficJam/TrafficJam/InputForm.cs
@@ -19,8 +19,8 @@ namespace TrafficJam
 
         private void btnStart_Click(object sender, EventArgs e)
         {
-            decimal percentageCars = percentageCarsNumeric.Value;
-            decimal maxSpeed = maxSpeedNumeric.Value;
+            double percentageCars = (double) percentageCarsNumeric.Value;
+            int maxSpeed = (int) maxSpeedNumeric.Value;
             uint countCars = (uint) countCarsNumeric.Value;
 
             var form = new MainForm(percentageCars, maxSpeed, countCars);
